requirements.txt
connexion==3.0.6
Flask==3.0.2
Flask_Login==0.6.3
flask_marshmallow==1.2.0
flask_sqlalchemy==3.1.1
python_bcrypt==0.3.2
Turbo_Flask==0.8.4
